third_party/blink/web_tests/virtual/select-parser-relaxation/external/wpt/html/semantics/forms/the-select-element/customizable-select/selectedcontent-restore.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] <selectedcontent> should be up to date after form restoration.
  assert_equals: selectedcontent.innerHTML should match the selected <option> expected "two<span>span</span>" but got ""
Harness: the test ran to completion.
